applies ESLint, Prettier
