+++ .vscode/launch.json
@@ -1,0 +1,25 @@
+{
+    "version": "0.2.0",
+    "configurations": [
+      {
+        "type": "node",
+        "request": "attach",
+        "name": "Next: Node",
+        "skipFiles": ["<node_internals>/**"],
+        "cwd": "${workspaceFolder}/src",
+        "port": 9229
+      },
+      {
+        "type": "chrome",
+        "request": "launch",
+        "name": "Next: Chrome",
+        "url": "http://localhost:3000",
+        "webRoot": "${workspaceFolder}/src",
+        "sourceMaps": true,
+        "sourceMapPathOverrides": {
+          "webpack://_N_E/*": "${webRoot}/*"
+        }
+      }
+    ]
+  }
+
