+++ vercel-nextjs-ts-app/.prettierrc
@@ -1,0 +1,5 @@
+{
+  "trailingComma": "es5",
+  "tabWidth": 2,
+  "arrowParens": "always"
+}
