--- vercel-nextjs-ts-app/pages/_app.tsx
@@ -1,8 +1,0 @@
-import '../styles/globals.css'
-import type { AppProps } from 'next/app'
-
-const MyApp: React.FC<AppProps> = ({ Component, pageProps }) => {
-  return <Component {...pageProps} />
-}
-
-export default MyApp
